UI for payment methods

+++ src/Plugins/Nop.Plugin.Payments.Manual/TransactMode.cs
@@ -3,7 +3,7 @@ namespace Nop.Plugin.Payments.Manual
     /// <summary>
     /// Represents manual payment processor transaction mode
     /// </summary>
-    public enum TransactMode : byte
+    public enum TransactMode : int
     {
         /// <summary>
         /// Pending
